Make creating a generic parse tree a form of action.
Previously the code to create generic parse trees was hard-coded into lrpar.
When we added code to also allow user-defined actions this ran alongside the
generic parse tree code (literally: the generic parse tree was always
created, even if you had user-specified actions!). The actions code was also
flawed in the sense that it gave incorrect results when errors were
encountered. This commit changes things in several ways.

First, we remove the hard-coded support for generic parse trees. Instead, we
view generic parse trees as just one thing you can create with actions. If
users want a generic parse tree, we create a "fake" action (the generic_ptree
function) which we attach to each production. [Note that this requires us
passing the RIdx to actions.]

Second, we remove the intermediate "action_vec" entirely. Instead we rely on
Vec::drain, which means that we only need "astack". When we call an action we
drain the last (pop_idx - 1) items and pass them to the action. Because I'm
lazy, we fully empty out the drain even for $x variables we don't use in
actions. That means that a grammar rules get translated to something along the
lines of:

  fn action(ridx: RIdx, input: &str, actions: Drain<...>) {
    let $1 = actions.next().unwrap();
    let $2 = actions.next().unwrap();
    ...
  }

Third, in order to keep the external API for generic parse trees simple, we
split Parser up into two impls: one for generic parse trees; one for actions;
the former (in essence) calls the latter. This does require threading ActionT
around more code than I'd like, but oh well.

Fourth, we add some basic tests to calc_actions, so that we can test action
code "for real".

Overall this commit is a useful simplification of the code, and gets us closer
to being correct too (user actions no longer crash by default when error
recovery happens). However, there are two eye-sores left. First, what to do with
the "input" string we pass around? That completely kills the hope of streaming
parsing, and it's also really awkward to deal with: this commit has a couple of
XXXs where I punted on dealing with this properly (the next commit will deal
with this). Second, how can user actions deal with repaired inputs? At the
moment they sometimes end up being passed empty strings, which cause calamity
(this requires a much more bigger fix).