Move away from the "store the whole input as a string" model.
This was starting to get baked in to too many parts of the system, which would
make removing it more difficult in the future.

This commit removes the storing of the whole input string idea. Instead,
the Lexer API now has a method `lexeme_str`. The only guarantee made is that if
you obtained a lexeme from `next()`, you can get the matching input substring
with `lexeme_str`. This will allow parsing of streaming data (you can start
parsing with incomplete data), although it does require users to keep all the
data they've streamed. Still, this is better than nothing and still keeps the
API fairly simple.

One day in the future, this API will allow us to truly lex on-the-fly, though
we'd need to buffer some lexemes (since error recovery, in a sense, "parses into
the future" before "returning to the present"). That's a bit tricky to do
nicely, and my two attempts to do so in a simple fashion quickly became bloated.
I'll worry about that in the future.