Force users to specify an action for every production.
If they don't, we panic at compile-time. This allows us to simplify the actions
code: we no longer need a Vec<Option<fn ...>> as Vec<fn ...> is sufficient.
Notice that we have to create a dummy action function for the start production
inserted by lrpar; to ensure it's never called its body is simply
"unreachable()".