Differentiate between lexemes of length 0 and lexemes created by error recovery.
The challenge here has long been that lexemes created by error recovery have
zero length. So is every zero length lexeme the result of error recovery?
Not necessarily: consider the DEDENT tokens in a Python program:

  if x:
     blah1
     if y:
        blah2
  z

there is only a single newline between "blah2" and "z" but 2 DEDENT lexemes, so
at least one of the DEDENT lexemes must be of length 0.

This commit gives the Lexeme type a way of saying "this lexeme has no length",
so that we can differentiate between lexemes of 0 length (e.g. DEDENT tokens)
and those lexemes with no length (that are the result of error recovery).

To prevent people shooting themselves in the foot, we also change the API so
that `length()` and `len()` return `Option` types; if they return None then the
lexeme has no length.