Fix parsing of %type declaration.
Previously this:

  %type T1 T2

was accepted and made the grammar's actiontype T2.

This commit changes the way this is parsed: basically anything after %type to
the end of line is now the type, so that the user can write things like:

  %type Result<u64, ()>

This commit also forbids repeated %type declarations, and adds appropriate
tests.