Instead of passing strings to action code, pass Result<Lexeme, Lexeme>.
This is two ideas squeezed into one.

First, passing strings for tokens destroys information: you can't get at the
line and column number, for example. Passing lexemes gives you access to such
information.

Second, some lexemes derive from user input, but some derive from error recovery
(and thus have no content). A good rule of thumb is that we don't care about
this difference for lexemes whose content we don't read (e.g. "filler" tokens
like "(" and ")" are there to help the parser and humans, but don't affect
evaluation), but we care deeply about it for lexemes whose contents we read
(e.g. identifiers, integers). Although the Lexeme API provides a way of
differentiating between the two, it's *really* easy to miss the difference. The
idea behind passing Result<Lexeme, Lexeme> is that the Ok case is for lexemes
that derive directly from user input and the Err case for lexemes generated by
error recovery. We thus force the user to think about this each time they want
to read a lexeme's content.

Fortunately, it's easy for users to come up with a sensible API that makes
dealing with the Result type easy. The calc_actions grammar, for example,
defines the actiontype thus:

  %actiontype Result<u64, ()>

When it's reading integer contents it does this:

      | 'INT' {
            let l = $1.map_err(|_| ())?;
            ...
         }

in other words, if Err(()) gets percolated upwards, then the calculator won't
evaluate a number. You can see this in practise with this session:

  >>> 1+2
  Result: 3
  >>> 1++2
  Parsing error at column 3.
  Result: 3
  >>> 1+18446744073709551616
  Error at column 3: '18446744073709551616' cannot be represented as a u64
  Unable to evaluate
  >>> 1+18446744073709551616+18446744073709551616
  Error at column 3: '18446744073709551616' cannot be represented as a u64
  Error at column 24: '18446744073709551616' cannot be represented as a u64
  Unable to evaluate

[My first attempt at this directly encoded the "action type must be a Result
type" idea into lrpar, before I realised that we can give the user the
flexibility to use this approach or not. Someone who's quickly hacking a parser
might be happy to panic!() as soon as they encounter a lexeme without content,
at which point the Result type would just be extra key presses for no good
reason.]