Change the parsing API's return results.
The API for getting results from parsing had become a nightmare to use. For
example the custom action parse method had this return type:

  Result<Node<{storaget}>, LexParseError<{storaget}, Node<{storaget}>>>

This looks innocuous, but required huge duplication to deal with correctly
e.g. this taken from calc_actions:

  match parse(&mut lexer) {
       Ok(Ok(v)) => println!("Result: {}", v),
       Ok(Err(_)) => println!("Unable to evaluate a result."),
       Err(LexParseError::LexError(e)) => // lexing error
       Err(LexParseError::ParseError(v, errs)) => {
           match v {
               Some(Ok(v)) => println!("Result: {}", v),
               _ => eprintln!("Unable to evaluate a result.")
           }
           for e in errs {
               // print parsing error
           }
       }
  }

Notice that we had to handle the success and failure conditions for the main
result twice. And, to add insult to injury, the API suggested that if things
went wrong, one either got a single lexing error, or one or more parsing errors.

This commit changes the return type to the simpler:

  (Option<Node<{storaget}>>, Vec<LexParseError<{storaget}>>)

One can achieve the same result as above with just:

  let (res, errs) = parse(&mut lexer);
  for e in errs {
  match e {
      LexParseError::LexError(e) => // print lexing error
      LexParseError::ParseError(e) => // print parsing error
  }
  match res {
      Some(Ok(r)) => println!("Result: {}", r),
      _ => println!("Unable to evaluate expression.")
  }

Notice also that LexParseError is now much simpler: it's just one lexing error,
or one parsing error (whereas before it was one lexing error or multiple parsing
errors). Even better, it's now clear that one can receive multiple lexing or
parsing errors, and that they can be intertwined. [OK, the current
implementation stops on the first lexing error, but at least that's just because
we're lazy, and not because the API forces us to!]