Add a pretty-printer on LexParseError.
Printing repair sequences out isn't rocket science but it's one more thing for
users to have to do when, in most cases, they'll probably be happy with the
"standard" format that nimbleparse produces.

This commit adds a "pp" function to LexParseError that returns a pretty-printed
string. Here's an example:

  Parsing error at line 1 column 9. Repair sequences found:
     1: Insert INT, Delete +
     2: Insert INT, Shift +, Insert INT

We have to pass the lexer to pp() so that it can put the user's input into the
pretty-printed string. Annoyingly, we also need to pass (one way or another)
%epp values to the pretty-printer. I grappled a bit with this. In one way, it
makes sense to pass the grammar, but that would lock us forever into statically
shipping the grammar (which we currently do, but which I wouldn't mind moving
away from one day). Instead we require a function which simply accepts TIDx's
and returns strings:

    pub fn pp<'a>(
        &self,
        lexer: &Lexer<StorageT>,
        epp: &Fn(TIdx<StorageT>) -> Option<&'a str>
    ) -> String;

To make life easier for users, CTBuilder now generates a function (called
token_epp):

  pub fn token_epp<'a>(tidx: TIdx<{storaget}>) -> Option<&'a str>

which fulfils this purpose. The implementation is simple and for the calculator
grammar looks like:

    const __GT_EPP: &[Option<&str>] = &[Some("PLUS"), Some("MUL"),
      Some("LBRACK"), Some("RBRACK"), Some("INT"), None];

    /// Return the %epp entry for token `tidx` (where `None` indicates "the token has no
    /// pretty-printed value"). Panics if `tidx` doesn't exist.
    #[allow(dead_code)]
    pub fn token_epp<'a>(tidx: TIdx<u8>) -> Option<&'a str> {
        __GT_EPP[usize::from(tidx)]
    }

so that in most cases you can just do this:

  pp(&lexer, &calc_y::token_epp)