Escape EPP strings when embedding them in string.
Previously this:

  %epp X "\"y\""

would lead to this code being generated:

  const __GT_EPP: &[Option<&str>] = &[Some(""y""), None];

which, of course, isn't valid:

  error: expected one of `)`, `,`, `.`, `?`, or an operator, found `""`

This commit escapes such strings before embedding them in source code.

It also adds a (currently bone-headedly) simple approach to properly testing
such things: it actually generates a complete project and runs "cargo build" on
it. This isn't quick (it takes well over a minute on my machine) so it's guarded
with #[ignore]. To run it execute "cargo test -- --ignored". This is enabled in
.travis.yml so that no-one can break this feature going forward, even though
they might only run "cargo test" locally.