Make the calc_actions tests part of ct.rs.
Previously calc_actions was an unhappy half-way house: it was both an example
for users, but also a way of testing compile-time code generation. Why were
there tests in an example? This seemed a recipe for future disaster if someone
removed the tests or (perhaps more likely) moved the examples to another
repository.

This commit moves those tests to ct.rs. They are harder to read in that format,
but it means that no-one can unintentionally remove them.