Add a %avoid_insert directive.
With the calc grammar and the input "2++3" there are two possible repair
sequences:

  Insert Int
  Delete +

If you're building a parse tree, both are equally reasonable. However, if you're
executing actions, "Insert Int" isn't great, because there is no value we can
insert, so we stop execution. In contrast, "Delete" allows us to continue
execution as normal. Since our current ranking is non-deterministic, that means
that we end up with this surprising outcome (from a real calc_actions session):

  >>> 2++3
  Parsing error at line 1 column 3. Repair sequences found:
     1: Insert INT
     2: Delete +
  Unable to evaluate expression.
  >>> 2++3
  Parsing error at line 1 column 3. Repair sequences found:
     1: Delete +
     2: Insert INT
  Result: 5

The problem is that, in general, we probably prefer to Insert new content than
delete content, on the basis that there is more value in what the user entered
than what they didn't enter. However, as the above example shows, tokens whose
value we care about are not ideal when recovering from errors. Unfortunately, we
have no idea which tokens have important values or not.

This commit adds a new grammar-level directive %avoid_insert which takes one or
more token names. When ranking repairs, repair sequences which contain an Insert
that references a token in %avoid_insert are penalised. If we add "%avoid_insert
'INT'" to the calc grammar, then it is guaranteed that the repair sequences
above are always returned in the following order:

  Delete +
  Insert Int