Make Rules an AST-level thing.
At some point, I hacked in the concept of ordered rules to the Yacc AST, but it
wasn't done in a very ergonomic way: in particular, there was no way of
attaching information to a rule (only to productions). This commit clears that
up: users now have to add a rule before they add a production. This gives us
more flexibility in the future.