Break the file generation code into several pieces.
process_file had become 312 lines of incomprehensible mess. This commit doesn't
make it perfect, but it breaks it up into several functions, the longest of
which is 109 lines. That makes it somewhat easier to look at a bit of code and
work out what it's part in the overall scheme is.