Allow different rules to have different action types.
This falls under the "why didn't this occur to me earlier?" heading.

Our previous %actiontype solution is, more-or-less, useless in practise. It
works in C because the union type that's created allows C programmers to treat
chunks of memory as dynamically typed (happily seg-faulting if the programmer
fished the wrong thing out of the union). Consider this (highly contrived)
grammar:

  %start S
  %%
  S: A | S A;
  A: 'a';

i.e. match one or more "a" characters. Let's assume that, for each "a" we match
we want to produce some value. How should we write this? Our only option is
"%actiontype Vec<...>" because "S" must return a Vec, even though "A" only
sensibly returns a singular value. In this case, this feels inelegant, but it
works.

However, what happens if you've got a real programming language building up an
AST? Consider a simple language which allow assignments and expressions:

  Assignments: Assignment | Assignments Assignment;
  Assignment: "ID" "=" Expr
  Expr: "INT" "+" Expr | "INT"

"Assignments" should return "Vec<Assign>" and "Expr" should return an "Expr". What should our %actiontype be now? Well, we can make an enum:

  enum AssignsOrExprs {
    Assigns(Vec<Assign>),
    Expr(Expr)
  }

and then we can have "%actiontype AssignsOrExprs". But we now have to
scale this up to every AST type in our entire program which is horrible, and
also turns our seemingly statically typed Rust into dynamically typed code: the
number of "match" statements doesn't bear thinking about!

How do other grammar generators deal with this? Well they do the obvious (in
retrospect) thing of allowing each different rule's actions to return different
types (note: each production in a rule must return the same type!). That's what
this commit does. Interestingly, this requires a simple tweak to cfgrammar; a
big change to ctbuilder.rs; and not much else.

First we add a YaccKind::Grmtools, a new variant on Yacc syntax. %actiontype is
not valid in this grammar type. Instead, each rule in this grammar type *must*
have a type after its name:

  %start S
  %%
  S -> Vec<A>: A { vec![$1] }
    | S A {
        $1.push($2);
        $1
    }
    ;
  A -> A: 'a' { A } ;
  %%
  pub struct A;

How does this work? Before we translated each production's action into a rule
which (for the first production in the grammar above) looked roughly like this:

  fn action_prod0(..., args: Drain<...>) -> Vec<A> {
     let arg1 = match args.next().unwrap() {
        AStackType::ActionType(x) => x,
        _ => unreachable!()
     };
     vec![arg1]
  }

The crucial detail now is that we know the types of all rules in advance. So we
generate an enum of all action types, and translate each action into a wrapper
and an action:

  enum ActionsKind {
    AKS(Vec<A>),
    AKA(A)
  }

  fn wrapper_prod0(..., args: Drain<...>) -> ActionsKind {
    let arg1 = match args.next().unwrap() {
      AStackType::ActionType(x) => x,
      _ => unreachable!()
    };
    ActionsKind::AKS(action_prod0(arg1))
  }

  fn action_prod0(..., mut arg1: A) -> Vec<A> {
    vec![arg1]
  }

  fn wrapper_prod1(..., args: Drain<...>) -> ActionsKind {
    let arg1 = match args.next().unwrap() {
      AStackType::ActionType(x) => x,
      _ => unreachable!()
    };
    let arg2 = match args.next().unwrap() {
      AStackType::ActionType(x) => x,
      _ => unreachable!()
    };
    ActionsKind::AKS(action_prod0(arg1, arg2))
  }

  fn action_prod1(..., mut arg1: Vec<A>, mut arg2: A) -> Vec<A> {
    arg1.push(arg2);
    arg1
  }

  fn wrapper_prod2(..., args: Drain<...>) -> ActionsKind {
    ActionsKind::AKA(action_prod2())
  }

  fn action_prod2(...) -> A {
    A
  }

The cunning thing about this is that we don't have to change the way the parser
works: it receives "action" functions that are wrappers, all of which have the
same function signature (i.e. from the parser's perspective it's a bit like we
wrote "%actiontype ActionsKind"). The wrapper functions then unpack the Drain,
and call the "actual" action functions which contain user code. The actual
action functions have their function's arguments and return types statically
typed, so Rust statically guarantees that you can't, for example, mix your
Classes and Imports. This does incur a mild additional overhead because of the
ActionsKind enum (one extra machine word, at worst, per ActionType we're
holding), but not enough to hugely worry us. And now we can write grammars which
generate ASTs with ease!

As a happy bonus, I realised that we can make arguments to action functions be
mutable (hence the "mut arg1: Vec<A>" above), which makes doing thing like
flattening lists a lot more ergonomic.