Make process_file consistent between LexerBuilder and CTParserBuilder
Previously, `process_file` in LexerBuilder expected a filename, while
CTParserBuilder's `process_file` expects a directory. This commit adjusts
LexerBuilder to generate the filename from a given directory similar to
CTParserBuilder, which makes the function a bit easier to use as can be
seen in lrpar/tests/build.rs where we the change saves a few lines of code.