Having offset_line_col return the byte offset of columns is annoying to use.
Byte offsets are efficient ways of indexing UTF-8 strings, but are confusing to
users, since bytes and characters aren't always the same thing. In other words,
in the face of multi-byte characters, offset_line_col gave confusing results.

This commit fixes this, returning the character offset for the column which is
nearly always what users want. To make this clear, this commit also renames
offset_line_col to line_col. This is a breaking change, but probably the right
thing to do.

[I'm not sure why this didn't occur to me earlier -- probably I made the classic
mistake of forgetting that byte and character offsets aren't the same thing in
UTF-8. Besides, before the surrounding_line function was added, there wasn't
really an easy way to perform the necessary calculation.]