Add the rule name to the generated Rust file.
This is a very simple way of helping debugging. Instead of getting a message
such as "error in gen_y.rs at line 2121" and then having to spend ages working
out which action that line of code relates to, this at least narrows it down to
a particular rule's productions.

In other words instead of:

    fn __gt_action_45(__gt_ridx: RIdx<u32>,
                     __gt_lexer: &Lexer<u32>,
    ...

one now gets:

    // DotOpt
    fn __gt_action_45(__gt_ridx: RIdx<u32>,
                     __gt_lexer: &Lexer<u32>,
    ...

It's a small difference, but it really helps. It would probably be nicer to do
something more here, perhaps even adding the full production text to the output.
That's quite a bit of work though.