Turn off some of Clippy's more dubious / opinionated warnings.
Clippy is a really good tool in many ways, and can really help improve code.
Unfortunately, some of its warnings are just bizarre. For example, it generally
bloats code to add a `Default` implementation, since it'll never be used: and
often there isn't even a sensible default to give!

This commit turns off several such dubious warnings found in grmtools packages.
I've only turned off the warnings in each package that occur in that package so
that we can reconsider this should it turn out to be a bad idea in the future.