Don't encode lexeme insertion in the API for lexemes end/length.
This seemed like a good idea when I created the API, but it's turned out to be a
pain in practise. The issue is that lexemes inserted by error recovery have no
contents or length. I was worried that users would forget this fact, and end up
in odd situations (e.g. having variables with names that are the empty string):
`end` and `len` both returned `None` if the lexeme was the result of error
recovery. In practise, this doesn't seem to be an issue: our actions API clearly
differentiates lexemes inserted by error recovery such that they don't really
leak into normal code. That makes the API verbose to use without much of a gain
with this idiom common:

  lexeme.end().unwrap_or_else(|| lexeme.start())

This commit thus simplifies things so that `end()` and `len()` return `usize`s
without concern for whether the lexeme is the result of error recovery or not.

Fixes #106.