Make CPCT+ the default error recovery algorithm.
On reflection, I think MF is too complex for the additional gains we get from
it over CPCT+. On that basis, this commit makes CPCT+ the default, and hides any
mention of MF from the outside world. At least for the time being, however,
we'll maintain all the code internally, in case we change our minds on this.