Fix two Clippy warnings and suppress two others.
The warnings are about unnecessary unwraps, but I've now tried 3 separate ways
of rewriting the code that offends Clippy and each time it becomes harder to
understand. I think the "unnecessary unwrap" warning is generally the right one,
but that the particular code fragments it complains about here are an unusual
exception: I've therefore suppressed the warnings only in the two functions in
question.