lifetime errors within generated code
Current issues appears to be that the external "'a" which is passed
to the generated "parse" function is being tied up with the shorter
liftimes which are passed `RTParserBuilder`.

So when `RTParserBuilder` is dropped at the end of the `parse` scope
the borrow checker assumes it's interior contents needs to live for
"'a" which is _at a minimum_ external of the generated "parse".

I'm not 100% sure what to do. But I'll play around with it tomorrow