Split the Lexer trait in two.
The previous `Lexer` trait was too inflexible: it could only be used with lexers
that were given their input in a single shot. This is fine for non-streaming
lexers, but if we ever want to support streaming lexers (e.g. that incrementally
lex input taken from a network) in the future, is clearly supoptimal. We might
as well bite the bullet now with the two traits looking as follows:

  pub trait Lexer<StorageT: Hash + PrimInt + Unsigned> {
    fn iter<'a>(&'a self) -> Box<dyn Iterator<Item = Result<Lexeme<StorageT>, LexError>> + 'a>;
  }

  pub trait NonStreamingLexer<'input, StorageT: Hash + PrimInt + Unsigned>: Lexer<StorageT> {
    fn span_str(&self, span: Span) -> &'input str;
    fn span_lines_str(&self, span: Span) -> &'input str;
    fn line_col(&self, span: Span) -> ((usize, usize), (usize, usize));
  }

In a theoretically perfect world, we would probably structure this as follows:

  pub trait Lexer<StorageT: Hash + PrimInt + Unsigned> {
    fn iter<'a>(&'a self) -> Box<dyn Iterator<Item = Result<Lexeme<StorageT>, LexError>> + 'a>;
  }

  pub trait LineColLexer<StorageT: Hash + PrimInt + Unsigned>: Lexer<StorageT> {
    fn line_col(&self, span: Span) -> ((usize, usize), (usize, usize));
  }

  pub trait NonStreamingLexer<'input, StorageT: Hash + PrimInt + Unsigned>: LineColLexer<StorageT> {
    fn span_str(&self, span: Span) -> &'input str;
    fn span_lines_str(&self, span: Span) -> &'input str;
  }

because one could build streaming lexers which can answer `line_col` queries.
However, splitting things into three traits is rather annoying, as most users
want access to `line_col` as well as `span_str`, so they would end up importing
`{LineColLexer, NonStreamingLexer}` which seems a bit of boilerplate too far. So
this commit collapses `LineColLexer` and `NonStreamingLexer` into one to avoid
this boilerplate.