Make it possible to customise the generated type name of a lexer.
By making use of the new LexerDef trait, we can make the compile-time builder
largely neutral about the type of lexer it's generating. All we have to do is,
somewhere, incorporate the type name (the new `LexerKind` enum tells us which
type name to use). There's probably a neater way of doing this than this commit
manages, but this already gives us a lot more flexibility for the future.