Admit defeat on being able to use custom rustfmt options.
rustfmt is a wonderful tool, but many of its options are unstable, and thus only
usable with nightly rustfmt. For a library that's aimed at Rust stable this is a
difficult to justify decision. This commit thus removes our rustfmt settings
entirely and reverts to rustfmt's defaults -- which means, mostly, lots and lots
and lots of trailing commas. Oh well.