Remove the pervasive assumption that the start state is 0.
There's no good reason for us exposing an assumption about which state in the
stategraph/statetable is the start state, and it may reduce our flexibility
later. This commit abstracts this assumption away such that the start state is
specified concretely just once (in pager::pager_stategraph), thus giving us a
single place to consider if we ever do want to change this assumption.