We no longer need to use the stategraph at runtime so remove that requirement.
The MF and KimYi recoverers required access to the stategraph. Since both have
now been removed, carting the stategraph around is particularly punishing for
compile-time use. The only thing that the parser requires of the stategraph is
knowledge of its start state, which we recently added to the statetable anyway,
so we no longer need the stategraph for parsing either. Even with a moderately
sized grammar such as yksom's, this removes 100KiB from the binary in release
mode (and 500KiB in debug mode!).

If we ever want to bring this feature back, we would probably do so in a more
efficient manner anyway (serialising the stategraph was hardly a long-term
solution).