Make the expression grammars left-associative.
I think -- but this is slightly lost to the hazy past -- that we based our
expression grammar on a right-associative grammar and perpetuated that
throughout grmtools. This commit makes all but one instance of the expression
grammar left-associative so that we do our best not to perpetuate
misunderstandings. The one remaining instance is the large stategraph and
statetable test: altering that would take considerable effort as it would need
to be carefully checked.