remove ParamT as a return value/some hacks for ctbuilder/tests to get it working.
The calc_parse_param.test and ctbuilder changes are fairly terrible hacks,
(Note that where it declares the type as u64, and then goes and uses it as an &mut u64).
But apparently this will work without return ParamT.

not sure if there is a way to avoid the library choosing a mutable ref for the caller, but it is a start.