depub grmtools.
This is a bit tricky, as grmtools intentionally exposes some details that no-one
currently uses, but they might one day (in particular in lrtable), but there are
also clearly some things that are "pub" that should never have been. I hope this
is a safe subset of the all the things that depub actually reported as being
possible to reduce in visibility.