Rename add_programs to set_programs.
`add_programs` suggests that you can repeatedly add new "programs" to an AST,
but calling the function repeatedly simply overrides whatever was there before.
`set_programs` is therefore a more appropriate name. However, since this is a
publicly facing API (though not one that many people are likely to use) we keep
the previous name around with a deprecation warning: at some point in the future
we can remove `add_programs` entirely.