Move to a traditional build() function that consumes the builder.
The previous `process` functions were unsatisfactory because they
mutated the `CTParserBuilder` (e.g. setting `conflicts`) but it was
never clear what would happen if you called them more than once. By
moving to a `build` function which consumes the builder and returns an
instance of `CTParser` we bypass this problem.

This change is still backwards compatible with previous code (with the
minor exception of the `conflicts` method moving from `CTParserBuilder`
to `CTParser`, but as an unstable API, that's arguably fair game).
However, I think the change is now just about annoying enough to make
that it feels unfair to put it in a compatible release: I think we'll
have to make the next release 0.11.0.