Generalise "inserted" lexemes to "faulty" lexemes.
At the moment the only type of "doesn't conform to definition" lexeme
one can have is a lexeme with length 0, but there's no reason to limit
that. For example, a language might only allow single-line strings and
thus the string in:

```
x = "blah
```

is clearly incorrect (it doesn't have a closing quote), but has a
non-zero length. At the moment our lexing API can't cope gracefully with
such a situation.

This commit thus tries to generalise the idea of an "inserted" lexeme to
a "faulty" lexeme: a faulty lexeme might have a non-zero length
(although the code doesn't currently support that!), and might even
exceed the length of its definition (although that's probably not going
to be very common).