Make Lexeme a trait, not a struct.
It is quite probable that in some parsing situations the way that
`Lexeme`s are represented in memory will be performance critical.
Previously we tried to satisfy everyone by having a fairly compact
`Lexeme` representation, but that meant that we couldn't represent some
reasonable size/lengths. I had hoped we could do something clever like
expand the `Lexeme` struct depending on `StorageT`s size, but that's not
possible in any sensible way.

This commit solves the problem in, I think, a more general way: it lets
users provide their own lexeme struct provided it conforms to the
(simple) `Lexeme` trait. This does mean lots of additional type
parameters, but (after three attempts!) these don't leak out to the end
user *too* much, though they do a bit.

As a half-way house, the "original" lexeme struct is now called
`StandardLexeme`. The plan is to move that out of lrpar and into lrlex,
but that can be done in a subsequent commit.

Interestingly, this commit paves the way for splitting lrlex and lrpar
apart in a more satisfactory way: the `Lexeme` trait makes clear that
any lexer can be used with lrpar. In other words, while lrlex only makes
sense with lrpar, the `Lexeme` trait will help make it clearer that
lrpar can be used without lrlex.