Move the concrete `StandardLexeme` struct to lrlex.
Having lrpar mandate use of a particular lexeme struct never felt
particularly satisfactory: this commit finally splits this so that lrpar
and lrlex are now much more obviously distinct. A vestigial lexeme
struct for testing-only purposes is contained in the new lrpar module
`test_utils` (for want of a better term), but it cannot be accessed by
external users.