Make it easier to create a lexer and parser in one go.
Many, probably even most, users will want to create a lexer and parser
in one go. lrpar's new "lexeme independent" flexibility comes at a
usability cost: you have to specify `LexemeT` at the very least
(`StorageT` can generally be inferred from that). That means that
there's more (literal and figurative) typing than I think most people
will want to bother with, at least when they're experimenting with
things.

This PR adds a relatively simple `lrpar_config` function to lrlex so
that if you want to use lrlex, which implicitly means using lrpar, you
have to do less faffing around. In essence instead of:

```rust
  let cp = CTParserBuilder::<StandardLexeme<u32>, _>::new()
      .yacckind(YaccKind::Grmtools)
      .grammar_in_src_dir("config.y")?
      .build()?;
  CTLexerBuilder::<u32>::new()
      .rule_ids_map(cp.lexeme_id_map())
```

you can say:

```rust
  CTLexerBuilder::new()
      .lrpar_config(|ctp| {
          ctp.yacckind(YaccKind::Grmtools)
              .grammar_in_src_dir("config.y")
              .unwrap()
      })
      .lexer_in_src_dir("config.l")?
      .build()?;
```

It's not an amazing improvement, but it does mean that most users don't
need to know about `StandardLexeme` and so on.

One annoying thing is that in order for this to type-check
automatically, one can't really rely on Rust's default type parameters,
as these only kick in when the type is unspecified (which doesn't really
happen in our setting). So this means that `CTLexerBuilder::new` has to
be simplified to "do the defaults but don't allow you to specify
`StorageT`" with a new `new_with_storaget` variant for when you do want
to specify `StorageT`. That's not ideal, but it means that those people
who care about bit-fiddling have a way of doing it, but those people who
don't care have a simple-ish API to deal with.