Make lrlex independent of lexeme type.
At some point we might want to introduce, say, `LexemePackedU8` for a
packed lexeme struct that fits in two machine words with a `u8`
`StorageT`, so it makes sense to make lrlex agnostic as to the
particular lexeme type used.

This commit does just that: `CTLexerBuilder::new` gives you a
`DefaultLexeme<u32>` as standard, but `CTLexerBuilder::new_with_lexemet`
allows you to choose precisely what type of lexeme (and what `StorageT`)
you want.