Make it possible for external callers to instantiate an LRNonStreamingLexer.
`lrpar` defines a pleasingly abstract `Lexer` and `NonStreamingLexer`
interface -- but that also means that common lexing tasks can be a bit
cumbersome. For example, if lrlex can't quite lex your language, you
currently have to do all the hard work of implementing `span_lines_str`
and friends. By making it possible for other users to instantiate
`LRNonStreamingLexer` we make it easier for people with simple lexing
tasks to avoid most of this boilerplate, while still allowing them to do
more exotic things with `lrpar`'s traits if they need to.