Add a Vec<Span> to YaccErrorKind::DuplicateAvoidInsert
The `DuplicateAvoidInsert` kind now contains the spans of all duplicates.
The span field of the `YaccError` has been changed to the location
of original `%avoid_insert` declaration which was duplicated.

In support of this, The GrammarAst `avoid_insert` field has been changed
from a `HashSet<String>` to a `HashMap<String, Span>`.