Add a Vec<Span> to YaccErrorKind::DuplicatePrecedence
The `DuplicatePrecedence` kind now contains the spans of all duplicates.
The span field of the `YaccError` has been changed to the location
of original precedence declaration which was duplicated.

In support of this, the `GrammarAst`'s `precs` field has been changed
from a `HashMap<String, Precedence>` to a `HashMap<String, (Precedence, Span)>`