Add a Vec<Span> to YaccErrorKind::DuplicateImplicitTokensDeclaration
The `DuplicateImplicitTokensDeclaration` kind now contains the spans
of all duplicates. The span field of the `YaccError` has been changed
to the location of the original `%implicit_tokens` declaration which
was duplicated.

In support of this, the `GrammarAst`'s `implicit_tokens` field
has been changed from `HashSet<String>` to `HashMap<String, Span>`.