Add a Vec<Span> to LexErrorKind::DuplicateName
The `DuplicateName` kind now contains the spans of all duplicates.
The span field of the `LexBuildError` has been changed to the location
of the original name which was duplicated.

Previously this error returned a span which was at the initial quotation
mark.  Now it points to the string inside the quotation mark.