Make StIdx use StorageT internally.
StIdxStorageT was previously hard-coded to u16: if you wanted to create
a grammar with more than 65536 states, you couldn't. This commit uses
StorageT in preference to StIdxStorageT.

In one sense this is icky, because now the user can't guess in advance
whether StorageT is big enough to store their statetable. However, the
other alternative was to add yet another type parameter, which seemed
even more icky.