Allow escaping of arbitrary characters beyond those required by regex.
In accordance with the POSIX lex specification, if a character is
not required to be escaped by the regex engine, it may still be escaped
within a regular expression, in which case the regex receives the
character itself. Existing escape sequences that were previously
accepted beyond those accepted by POSIX lex should still pass through
to regex as before.

The primary motivation for this change is to allow escaping an intial
`<` character in a regular expression, allowing it to be differentiated
from the beginning of start state.